libs/strings: Move into internal
