+++ internal/strings/string.go
